% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpplot.R
\name{lpplot}
\alias{lpplot}
\title{Linear Programming Solutions
`lpplot()` gives the minimum/maximised value and feasible region for LP problems}
\usage{
lpplot(vec, obj, crit = "max")
}
\arguments{
\item{crit}{}
}
\description{
Linear Programming Solutions
`lpplot()` gives the minimum/maximised value and feasible region for LP problems
}
